requirements.txt
termcolor==3.1.0
pydantic==2.11.5
google-genai>=1.40.0
playwright==1.52.0
browserbase==1.3.0
browser-use==0.8.1
langchain==0.3.27
langchain-community==0.3.31
langchain-experimental==0.3.4
langchain-openai==0.3.35
langchain_google_genai==2.1.12
nest-asyncio==1.6.0
rich
pytest
transformers
qwen-vl-utils
qwen-agent
openai>=1.68.2
pyautogui>=0.9.54
Pillow>=11.1.0
torch
torchvision
numpy==1.26.4
pandas==2.2.2
streamlit
anthropic